pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "qrcraft"
version = "1.0.0"
description = "Beautiful, animated QR codes with gradient fills, custom finder patterns, circular modules, and logo overlays."
readme = "README.md"
license = { text = "MIT" }
requires-python = ">=3.10"
authors = [
    { name = "Salus Financial Technology", email = "engineering@salusfintech.com" },
]
keywords = ["qr-code", "qr", "gradient", "animated", "branding", "cli"]
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Developers",
    "Intended Audience :: End Users/Desktop",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Programming Language :: Python :: 3.13",
    "Topic :: Multimedia :: Graphics",
    "Topic :: Utilities",
]
dependencies = [
    "qrcode[pil]>=7.4",
    "Pillow>=10.0",
]

[project.urls]
Homepage = "https://github.com/salusfintech/qrcraft"
Repository = "https://github.com/salusfintech/qrcraft"
Issues = "https://github.com/salusfintech/qrcraft/issues"

[project.scripts]
qrcraft = "qrcraft.cli:main"

[tool.hatch.build.targets.wheel]
packages = ["src/qrcraft"]
